requirements.txt
python-dateutil
DBUtils==1.3
Flask==1.0.2
numpy
pandas
PyMySQL==0.9.2
requests==2.18.4
Werkzeug==0.14.1
flask-cors
xlwt
xlrd
xlutils
flask_restful
python-docx
PyJWT
wtforms